requirements.txt
requests>=2.25.0
Pillow>=10.0.0
python-dotenv>=1.0.0